+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3626,7 +3626,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KT2Z4NT89N8Q8WFPR5T406T1"),
+  id: z.string().default("01KT376M07FVZX3GTZYS84AB8P"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),